@@ -22,8 +22,9 @@
         "silverstripe/sentry": "*"
     },
     "require": {
-        "php": "^5.4|<7.5",
-        "sentry/sdk": "^2",
+        "php": "^7.4.0|^8",
+        "sentry/sdk": "^3",
+        "composer/package-versions-deprecated": "^1.11",
         "silverstripe/framework": "^4"
     },
     "require-dev": {
@@ -36,7 +37,8 @@
         }
     },
     "config": {
-        "process-timeout": 900
+        "process-timeout": 900,
+        "optimize-autoloader": true
     },
     "extra": {
         "installer-name": "sentry"
