@@ -9,6 +9,10 @@
 
 namespace PhpTek\Sentry\Log;
 
+use Composer\InstalledVersions;
+use Monolog\Logger;
+use Sentry\Frame;
+use Sentry\Client;
 use SilverStripe\Control\Director;
 use SilverStripe\Control\Middleware\TrustedProxyMiddleware;
 use SilverStripe\Core\Injector\Injector;
@@ -35,51 +39,69 @@ class SentryLogger
      * Stipulates what gets shown in the Sentry UI, should some metric not be
      * available for any reason.
      *
-     * @const string
+     * @var string
      */
     const SLW_NOOP = 'Unavailable';
 
     /**
-     * A static constructor as per {@link Zend_Log_FactoryInterface}.
+     * Default text to show if in self-generated stacktraces, we're unable to discern data.
+     *
+     * @var string
+     */
+    public const DEFAULT_FRAME_VAL = 'Unknown';
+
+    /**
+     * Factory, consumed by {@link SentryHandler}.
      *
      * @param  array $config    An array of optional additional configuration for
      *                          passing custom information to Sentry. See the README
      *                          for more detail.
+     * @param  Client $client
      * @return SentryLogger
      */
-    public static function factory(array $config = []) : SentryLogger
+    public static function factory(array $config = [], Client $client): SentryLogger
     {
-        $env = $config['env'] ?? [];
-        $user = $config['user'] ?? [];
-        $tags = $config['tags'] ?? [];
-        $extra = $config['extra'] ?? [];
-        $level = $config['level'] ?? [];
-        $logger = Injector::inst()->create(static::class);
+        $logger = new static();
+        $env = $logger->defaultEnv();
+        $tags = $logger->defaultTags();
+        $extra = $logger->defaultExtra();
+        $user = [];
+        $level = Logger::DEBUG;
 
-        // Set default environment
-        $env = $env ?: $logger->defaultEnv();
-        // Set any available user data
-        $user = $user ?: [];
-        // Set any available tags available in SS config
-        $tags = array_merge($logger->defaultTags(), $tags);
-        // Set any available additional (extra) data
-        $extra = array_merge($logger->defaultExtra(), $extra);
+        if ($config) {
+            $env = $config['env'] ?? $level;
+            $tags = array_merge($tags, $config['tags'] ?? []);
+            $extra = array_merge($extra, $config['extra'] ?? []);
+            $user = $config['user'] ?? $user;
+            $level = $config['level'] ?? $level;
+        }
 
-        $logger->adaptor->setContext('env', $env);
-        $logger->adaptor->setContext('tags', $tags);
-        $logger->adaptor->setContext('extra', $extra);
-        $logger->adaptor->setContext('user', $user);
-        $logger->adaptor->setContext('level', $level);
+        $adaptor = (new SentryAdaptor($client))
+            ->setContext('env', $env)
+            ->setContext('tags', $tags)
+            ->setContext('extra', $extra)
+            ->setContext('user', $user);
 
-        return $logger;
+        return $logger->setAdaptor($adaptor);
     }
 
     /**
      * @return SentryAdaptor
      */
-    public function getAdaptor() : SentryAdaptor
+    public function getAdaptor(): SentryAdaptor
     {
         return $this->adaptor;
+    }
+
+    /**
+     * @param  SentryAdaptor $adaptor
+     * @return SentryLogger
+     */
+    public function setAdaptor(SentryAdaptor $adaptor): SentryLogger
+    {
+        $this->adaptor = $adaptor;
+
+        return $this;
     }
 
     /**
@@ -88,7 +110,7 @@ class SentryLogger
      *
      * @return string
      */
-    public function defaultEnv() : string
+    public function defaultEnv(): string
     {
         return Director::get_environment_type();
     }
@@ -108,13 +130,15 @@ class SentryLogger
      *
      * @return array
      */
-    public function defaultTags() : array
+    public function defaultTags(): array
     {
         return [
-            'Request-Method'=> $this->getReqMethod(),
-            'Request-Type'  => $this->getRequestType(),
-            'SAPI'          => $this->getSAPI(),
-            'SS-Version'    => $this->getPackageInfo('silverstripe/framework')
+            'request.method'=> self::get_req_method(),
+            'request.type' => self::get_req_type(),
+            'php.sapi' => self::get_sapi(),
+            'silverstripe.framework.version' => self::get_package_info('silverstripe/framework'),
+            'phptek.sentry.version' => self::get_package_info('phptek/sentry'),
+            'app' => self::get_app_info(),
         ];
     }
 
@@ -126,36 +150,11 @@ class SentryLogger
      *
      * @return array
      */
-    public function defaultExtra() : array
+    public function defaultExtra(): array
     {
         return [
-            'Peak-Memory'   => $this->getPeakMemory()
+            'PHP Peak Memory' => self::get_peak_memory(),
         ];
-    }
-
-    /**
-     * Return the version of $pkg taken from composer.lock.
-     *
-     * @param  string $pkg e.g. "silverstripe/framework"
-     * @return string
-     */
-    public function getPackageInfo(string $pkg) : string
-    {
-        $lockFileJSON = BASE_PATH . '/composer.lock';
-
-        if (!file_exists($lockFileJSON) || !is_readable($lockFileJSON)) {
-            return self::SLW_NOOP;
-        }
-
-        $lockFileData = json_decode(file_get_contents($lockFileJSON), true);
-
-        foreach ($lockFileData['packages'] as $package) {
-            if ($package['name'] === $pkg) {
-                return $package['version'];
-            }
-        }
-
-        return self::SLW_NOOP;
     }
 
     /**
@@ -164,9 +163,9 @@ class SentryLogger
      *
      * @return string
      */
-    public function getRequestType() : string
+    public static function get_req_type(): string
     {
-        $isCLI = $this->getSAPI() !== 'cli';
+        $isCLI = self::get_sapi() !== 'cli';
         $isAjax = Director::is_ajax();
 
         return $isCLI && $isAjax ? 'AJAX' : 'Non-Ajax';
@@ -177,7 +176,7 @@ class SentryLogger
      *
      * @return string
      */
-    public function getPeakMemory() : string
+    public static function get_peak_memory(): string
     {
         $peak = memory_get_peak_usage(true) / 1024 / 1024;
 
@@ -189,15 +188,9 @@ class SentryLogger
      *
      * @return string
      */
-    public function getUserAgent() : string
+    public function getUserAgent(): string
     {
-        $ua = @$_SERVER['HTTP_USER_AGENT'];
-
-        if (!empty($ua)) {
-            return $ua;
-        }
-
-        return self::SLW_NOOP;
+        return $_SERVER['HTTP_USER_AGENT'] ?? self::SLW_NOOP;
     }
 
     /**
@@ -205,23 +198,52 @@ class SentryLogger
      *
      * @return string
      */
-    public function getReqMethod() : string
+    public static function get_req_method(): string
     {
-        $method = @$_SERVER['REQUEST_METHOD'];
-
-        if (!empty($method)) {
-            return $method;
-        }
-
-        return self::SLW_NOOP;
+        return $_SERVER['REQUEST_METHOD'] ?? self::SLW_NOOP;
     }
 
     /**
      * @return string
      */
-    public function getSAPI() : string
+    public static function get_sapi(): string
     {
         return php_sapi_name();
+    }
+
+    /**
+     * @param  string $package
+     * @return string
+     */
+    public static function get_package_info(string $package): string
+    {
+        return InstalledVersions::getVersion(trim($package)) ?? self::SLW_NOOP;
+    }
+
+    /**
+     * Format and return a string of metadata about the app in which this module is installed.
+     *
+     * @return string
+     */
+    public static function get_app_info(): string
+    {
+        $meta = InstalledVersions::getRootPackage();
+        $data = [
+            'project' => $meta['name'] ?? self::SLW_NOOP,
+            'branch' => $meta['version'] ?? self::SLW_NOOP,
+            'commit' => $meta['reference'] ?? self::SLW_NOOP,
+        ];
+
+        // If we have nothing to show, then show a meaningful default
+        $filtered = array_filter($data, function ($item): bool {
+            return $item == self::SLW_NOOP;
+        });
+
+        if (count($filtered) === count($data)) {
+            return self::SLW_NOOP;
+        }
+
+        return sprintf('%s: %s (%s)', $data['project'], $data['branch'], $data['commit']);
     }
 
     /**
@@ -230,7 +252,7 @@ class SentryLogger
      *
      * @return string
      */
-    public static function get_ip() : string
+    public static function get_ip(): string
     {
         $headerOverrideIP = null;
 
@@ -273,14 +295,14 @@ class SentryLogger
      * @param  mixed Member|null $member
      * @return array
      */
-    public static function user_data(Member $member = null) : array
+    public static function user_data(Member $member = null): array
     {
         if (!$member) {
             $member = Security::getCurrentUser();
         }
 
         return [
-            'IPAddress' => self::get_ip() ?: self::SLW_NOOP,
+            'IP Address' => self::get_ip() ?: self::SLW_NOOP,
             'ID'       => $member ? $member->getField('ID') : self::SLW_NOOP,
             'Email'    => $member ? $member->getField('Email') : self::SLW_NOOP,
         ];
@@ -290,15 +312,19 @@ class SentryLogger
      * Manually extract or generate a suitable backtrace. This is especially useful
      * in non-exception reports such as those that use Sentry\Client::captureMessage().
      *
+     * Note: Calling logic only makes use of this if the "custom_stacktrace" option is enabled.
+     * As it is, it requires a little more work to make it on-par with Sentry's defaults (which is this
+     * module's default also).
+     *
      * @param  array $record
-     * @return array
+     * @return array An array of filtered Sentry\Frame objects.
      */
-    public static function backtrace(array $record) : array
+    public static function backtrace(array $record): array
     {
         if (!empty($record['context']['trace'])) {
             // Provided trace
             $bt = $record['context']['trace'];
-        } elseif (isset($record['context']['exception'])) {
+        } else if (isset($record['context']['exception'])) {
             // Generate trace from exception
             $bt = $record['context']['exception']->getTrace();
         } else {
@@ -307,8 +333,8 @@ class SentryLogger
 
             // Push current line into context
             array_unshift($bt, [
-                'file'     => $bt['file'] ?? 'N/A',
-                'line'     => $bt['line'] ?? 'N/A',
+                'file'     => $bt['file'] ?? self::DEFAULT_FRAME_VAL,
+                'line'     => (int) ($bt['line'] ?? 0),
                 'function' => '',
                 'class'    => '',
                 'type'     => '',
@@ -317,15 +343,29 @@ class SentryLogger
         }
 
         // Regardless of where it came from, filter the exception
-        return Backtrace::filter_backtrace($bt, [
+        $filtered = Backtrace::filter_backtrace($bt, [
             '',
             'Monolog\\Handler\\AbstractProcessingHandler->handle',
             'Monolog\\Logger->addRecord',
             'Monolog\\Logger->error',
             'Monolog\\Logger->log',
             'Monolog\\Logger->warn',
-            'PhpTek\\Sentry\\Handler\\SentryHandler->write',
-            'PhpTek\\Sentry\\Handler\\SentryHandler->backtrace',
+            'PhpTek\\Sentry\\Handler\\SentryHandler',
+            'PhpTek\\Sentry\\Handler\\SentryHandler::write',
+            'PhpTek\\Sentry\\Logger\\SentryLogger::backtrace',
         ]);
+
+        return array_map(function($item) {
+            return new Frame(
+                $item['function'] ?? self::DEFAULT_FRAME_VAL,
+                $item['file'] ?? self::DEFAULT_FRAME_VAL,
+                $item['line'] ?? 0,
+                $item['raw_function'] ?? self::DEFAULT_FRAME_VAL,
+                $item['abs_filepath'] ?? self::DEFAULT_FRAME_VAL,
+                $item['vars'] ?? [],
+                $item['in_app'] ?? true,
+            );
+        }, $filtered);
     }
+
 }
